refactored Map to separate layers/markers
